scripts: remove /debug_ramdisk requirement
theres no need nowadays for most modules to have this unless youre on
magic mount. on magic mount this is needed for workdir.

but then if youre on .nomount, the tmpfs is useless so you can just turn the damn thing off.

most modules nowadays are now adapted to missing /debug_ramdisk tmpfs.
figured out this modules should start moving to this direction too.

please review and test before merging.
another is that I am not planning to check webui side of things.